==> News_Articles/in_tibet_website/sertha/1_all_categories.ipynb <==
from bs4 import BeautifulSoup
import requests
import re


# url = "https://sertha.net/"

# headers = {
#         'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
#     }
# response = requests.get(url, headers=headers)
# response.raise_for_status()

# # Parse the page content with BeautifulSoup
# soup = BeautifulSoup(response.content, 'html.parser')
# all_menu = soup.find_all("li", class_="menu-item menu-item-type-taxonomy menu-item-object-category menu-item-has-children")
# all_categories = []
# print(all_menu)
# if all_menu:
#     for each_category in all_menu:
#         print(each_category)
#         each_link = each_category.find("a", class_="elementor-item has-submenu")
#         if each_link:
#             href = each_link.get('href')
#             if href:
#                     text = link.get_text(strip=True)
#                     all_categories.append((text, href))

issues with extraction:
- doing manuly

all_categories = []

all_categories.append(["https://sertha.net/category/hhdl/", "༧གོང་ས་མཆོག"])
all_categories.append(["https://sertha.net/category/%e0%bd%a2%e0%be%a9%e0%bd%bc%e0%bd%98%e0%bc%8b%e0%bd%a2%e0%bd%b2%e0%bd%82/", "རྩོམ་རིག"])
all_categories.append(["https://sertha.net/category/%e0%bd%a6%e0%be%b3%e0%bd%bc%e0%bd%96%e0%bc%8b%e0%bd%82%e0%bd%a6%e0%bd%bc%e0%bc%8d/", "སློབ་གསོ།"])
all_categories.append(["https://sertha.net/category/%e0%bd%82%e0%bd%a6%e0%bd%a2%e0%bc%8b%e0%bd%a2%e0%be%a9%e0%bd%bc%e0%bd%98%e0%bc%8d/", "གསར་རྩོམ།"])
all_categories.append(["https://sertha.net/category/%e0%bd%a3%e0%bd%bc%e0%bc%8b%e0%bd%a2%e0%be%92%e0%be%b1%e0%bd%b4%e0%bd%a6%e0%bc%8d/", "ལོ་རྒྱུས།"])
all_categories.append(["https://sertha.net/category/%e0%bd%82%e0%bd%a6%e0%bd%a2%e0%bc%8b%e0%bd%a0%e0%bd%82%e0%be%b1%e0%bd%b4%e0%bd%a2%e0%bc%8d/", "གསར་འགྱུར།"])
all_categories.append(["https://sertha.net/category/%e0%bd%98%e0%bd%b2%e0%bc%8b%e0%bd%a6%e0%be%a3%e0%bc%8b%e0%bd%84%e0%bd%bc%e0%bc%8b%e0%bd%a6%e0%be%a4%e0%be%b2%e0%bd%bc%e0%bd%91%e0%bc%8d/", "མི་སྣ་ངོ་སྤྲོད།"])
all_categories.append(["https://sertha.net/category/%e0%bd%a6%e0%be%b2%e0%bd%bc%e0%bd%a3%e0%bc%8b%e0%bd%a2%e0%be%92%e0%be%b1%e0%bd%b4%e0%bd%93%e0%bc%8d/", "སྲོལ་རྒྱུན།"])


all_categories

len(all_categories)

## Extarct each categories page number

def get_page_no(url):
    """
    
    """

    headers = {
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
    }
    try:
        # Make the request to the URL
        response = requests.get(url, headers=headers)
        response.raise_for_status()
        
        # Parse the page content with BeautifulSoup
        soup = BeautifulSoup(response.content, 'html.parser')
        page_content = soup.find("nav", class_="elementor-pagination")
        # print(page_content)
        if page_content:
            
            all_page_link = page_content.find_all("a", class_="page-numbers")
            if all_page_link:
                # print(all_page_link[-2])
                page_no = all_page_link[-2].get_text(strip=True)
                page_no = re.search(r'Page(\d+)', page_no)
                if page_no:
                    return [all_page_link[-2].get("href"), page_no.group(1)]
                
        return ["no link last page link" , 0]
    except Exception as e:
        print(e)
        print(url, "\n\n")
        return ["no link last page link" , 0]

url = "https://sertha.net/category/hhdl/"
# url = "https://ti.zangdiyg.com/category/index/id/134/p/2.html"

get_page_no(url)


final_link = []
new_categories = all_categories
i = 0
for each_link in new_categories:
    link = each_link[0]
    page_no = get_page_no(link)
    if "no link last page link" == page_no[0]:
        print(page_no)
        
    clean_list = []
    clean_list.extend(each_link)
    clean_list.extend(page_no)
    # print(clean_list)
    final_link.append(clean_list)
    i = i+1
    if i%5 == 0:
        print(i)


final_link[:5]

## saving the final file

import pandas as pd

df = pd.DataFrame(final_link, columns=[ 'original_url', 'name_of_cate', 'Last_page_link', 'list_page_no'])
df.head()

df.head(30)

df.to_csv("./data/All_cate.csv", index=False)

df.shape
